dx: add support for inverting all types of cursor
This commit introduces support for inverting all types of cursor.

To invert the display without re-rendering any text, we draw the cursor
into a command list and then compose the command list with the existing
renderer using the MASK_XOR composition flag.

This wouldn't normally work with our renderer because there is no
_background_ color to invert in some cases (such as when acrylic is in
use.)

To work around that, we're taking advantage of @zadjii-msft's two-pass
cursor renderer: First, we draw the cursor region in the background
color (to give the later pass something to work with), then after the
text is drawn we do the command list MASK_XOR.

Example cursor draw (for emptyBox cursor)
Therefore, cursor drawing looks like this (`_` is background, `@` is
cursor, `A` is glyph)

```
First pass (before text)

_____
_   _
_   _
_   _
_____

Text is added

__A__
_A A_
_AAA_
_A A_
_A_A_

Second pass (after text)

@@a@@ < Drawn with XOR mask; overlaps shown as [a]
@A A@
@AAA@
@A A@
@a@a@
```

The box, underline, etc. cursors are treated the same way, give or take
some pixels.

Fixes #9610